feat: runInWorkerUntilTrue : ensure bridge is ready before running one more time
Sometimes, when the connection is slow or the website is slow,
runInWorkerUntilTrue would retry running the function before the bridge
is ready. This caused error like '*** method is not implemented'.

Now, an event is emitted when the bridge is ready, and the launcher can
listen to it to know it can run the method once again and be sure the
bridge is ready.