feat: add waitForElementInWorker in ContentScript
Now the ContentScript class has a waitForElementInWorker method which
will wait for a dom element, selected by a css selector, to be present
in the worker's dom. This method will work even if the worker current
page is reloaded, via redirect or user's action

This method will avoid many synchronization problems in connectors.