feat: add connector installation for github repository
When the user runs a client side connector from harvest, the source code
of the connector is fetched from the corresponding github repository
and extracted before being run.

Since github sources are for development purposes, no cache is applied
to these and the content script code is fetched each time the connecor
is run.

Registry source will be added later