feat: Let's display the Cozy Home when running on iOS
Local webviews are a mess. It doesn't work the same way between iOS &
Android.

On iOS we need to read the content of the file to get the HTML and set
the baseDir.

The code needs to be tested on Android.

Also, it's only the step since we can't interact with the Home right now
Clicks don't trigger any change for now.