feat: Inject location.hash in CozyWebView when url has one
When the HomeView injects `connected/<connectorId>` in the url for the
first time, as the URL is new then CozyWebView is rendered correctly
with the Connectors pane open

But when HomeView injects it for the second time, if `connectorId` is
the same, then the URL is not considered as changed by CozyWebView and
it is not re-rendered and the Connectors pane does not open

If we inject `location.hash` when a hash is detected, then this problem
is fixed

However this method is not optimal as a re-render of HomeView would
trigger the hash update and reopen the Connectors pane. A better
solution would be to use a cozy-intent native-to-native message. We
should implement this in the future