feat: Change page reload detection method
We now detect page reload with a postMessage sent right
at the beginning of the contentent script. Since the content script
is reloaded on every page in the main frame, the postMessage event will
be sent when the page is really reloaded, avoiding to filter, redirects,
multiple reloads of the same page on ios etc.

This makes running ccc on ios and android more stable