feat: Implement UI transition to the Authorize view
This commit implement transition when entering the authorize view

This implies to split the OAuth handling in two separate methods

`callInitClient` now only manage the client connection and the new
`authorizeClient` method will handle the authorize part

We expected to create a fluid transition to the InAppBrowser's
authorization page but the used library does not allow to pre-load the
InAppBrowser before displaying it. So current transition still shows a
some loading when entering the InAppBrowser. This behavior may be
improved in the future