fix(transition): Set transition targetSize to smaller size
When transitionning to the Authorization view, the Cozy logo is shown
and animated to grow and fill the screen

On Android with a big `display size` setting, this transition may crash
with a "too large bitmap" error message

In our case the dimension of the canvas, managed by SvgView from the
library React-Native-SVG, becomes bigger than 5000 pixel equivalent

This is because the animation used a raw pixel size that was dependent
of the device's pixel ratio. When increasing the Android's display size
then the pixel ratio also increase

This makes the Cozy logo increase in size as the pixel ratio is applied
after the image size is set from the React component

That's why we set a Animated.View height and width to a size that is 2
times the window's height. We ensure that the logo will grow to fill
the entire screen, but as the window's height is relative to the pixel
ratio we expect to get a reasonable value

However this may not work on all devices as some devices have a very
large pixel ratio. But those devices tend to have more memory allocated
for diplaying large images

More details: https://reactnative.dev/docs/pixelratio