feat: Upgrade React Native Webview
Some Client Side Konnectors (Clisk) are
receiving a blob url and then try to
download the file. The previous version
of RNW was crashing the app since the
blob url was sended to the downloadManager
and then the downloadManager will crash
because it doesn't handle this kind of url
(aka blob://https://...). It only handles
http(s):// url.

This is the needed fix:
https://github.com/react-native-webview/react-native-webview/pull/2328

I didn't try to run the app on iOS yet
with this upgrade. This is currently only
a branch for testing purpose.