feat: Do not try to intercept `index.html` if no local bundle exists
In current implementation, `index.html` interception mechanism is used
on HomeView in order to use local assets instead of cozy-stack ones

This interception always happens as local assets are always available.
They either exists from embedded assets or from previously downloaded
cozy-stack's assets

In following commit we will want to use this mechanism for other
cozy-apps

As we expect to embed assets in app's binaries only for cozy-home, then
we cannot ensure that assets are always available for other cozy-apps.
They are not available until we download them from cozy-stack

So if no local assets are available, we want to load the cozy-app
directly from the cozy-stack without any proxy