fix: Don't use CSS to add the safe-area-inset-bottom
We need to be clear on where is the responsibility to handle the
SafeArea.

Sometimes this is handled within the HTML, sometimes this is handled by
the native side, and sometimes by the both. The result is that the
bouton at the bottom of the page is never exactly at the same place.