feat: Add plugin for managing AsyncStorage
ATM. It's hard to know what informations are the AsyncStorage.
 So let's add a few flipper plugins to see what's going on. I added
two plugins, because one let us edit the content of the async storage,
but the display is not good, the other has a better display but don't
let us edit the content.

You need to install async-storage and async-storage-advanced within
the flipper plugins window.

Those plugins are only enabled on DEV mode. Not in release mode.