fix: Konnector's WebView Media behavior
On Android, by default, the webview will require
an user itereaction to play a video with the
autoplay attribute.

By default, on iOS, the webview will play the
"autoplay" video. Even if the fullscreen mode
is true.

It resulted that, when we run the hidden content
script, sometimes video has been played in
fullscreen.

By requesting the user interaction on the media
we avoid this issue and save bandwidth to the user.