fix: Viewport size on iOS
We randomly have isMobile equal to
false on iOS. It seems that the app's
script can be executed before having
the dom available. So when runing
the useBreakpoint() the value can be
the one by default: aka 980. Then
the viewport is changed to 390
(base on the meta viewport). But when
the viewport changed, the resize method
is not called.

https://stackoverflow.com/questions/5508455/mobile-safari-window-reports-980px/35987682#35987682

So, to get the right value, we dispatch
a resize event when the DOM is loaded
in order to wakeup UI's breakpoint.