fix: Remount WebView when terminated by iOS memory management
In some scenario iOS may terminate the WKWebView process

This can happen when the app is running in background for a long time

When this happens, the WebView component becomes blank

SupervisedWebView is a component that should mitigate this behavior by
remounting the WebView after it has been terminated by the OS

Note that the official react-native-webview documentation uses
`webview.reload()` to do the reload part. But this is not working when
using the `HttpServer` (with html injection). So instead we remount the
entire component

More info: https://nevermeant.dev/handling-blank-wkwebviews/