fix: Handle capabilities as object
Capabilites suffered the same issues as flag object.
Due to multiple stringify and the fact these values are object,
there was a mistake where there were scoped into quotes in the end json
This is not valid JSON so it made the whole app bootstrap crash.
parsing/stringifying JSON is inherently risky, what should we do,
except logging, when an error arises here?