fix: Delete existing version folder if exists when updating a cozy-app
In current implementation we have a mechanism that tries to update the
cozy-app local bundle from stack, but skips the package download if it
already exists in local folder

This has been implemented to optimise recurrent versions switches when
going back and forth between prod and dev versions

But this was a bad idea as this scenario is pretty rare and it would
not fix existing corrupted bundles from previously aborted extraction

This can occurs when the user kills the application at the exact time
the app is extracting the downloaded bundle. In this case we have a
partial bundle in local folder, but the version is not updated. Then on
next start, the app tries to update the bundle, detects that a local
folder already exist and so skip the download phase. The result makes
the app unusable as it tries to display the cozy-app using a corrupted
local bundle

With new implementation, we chose to delete the existing folder and
always download the bundle, so the described scenario cannot happen
anymore