fix: Don't intercept cozy download links
On Android, download links are functional and are handled automatically
by the webView

With current implementation, those are intercepted by
`interceptNavigation` as they are considered as outgoing links. So
nothing happens when clicking on download links

Instead we want `interceptNavigation` to let the navigation happen

On iOS, download links are disabled until we correctly handle downloads
from the iOS webView. So this commit should not impact iOS yet