fix: Disable `universalApk` on Release build
When building a release and since we activated
`enableSeparateBuildPerCPUArchitecture`, then the
`output.versionCodeOverride` variable gets a different value for each
CPU architecture. This adds a number suffix to the project's version
number

For example if the project's version number is `2401`, then we would
get versions `2401001` for `armeabi-v7a`, `2401002` for `x86` etc.

The generate `.aab` file seems to get the last architecture version
number as it's global version number

You can check a version number for the generated `.aab` file by
executing
```
brew install bundletool
bundletool dump manifest --bundle app-release.aab
```

When `universalApk` is true, the last generated architecture seems to
be the universal one, which doesn't override the version number. So in
this case the generated `.aab` version number would be `2401`

Since we published a version without `universalApk` in the past, the
higher version number in the PlayStore is `2301004`, so we cannot
publish a version like `2401` anymore

So we want to disable `universalApk` for future releases to force
`240100X` format