fix: Do not force nav and status bar colors on iOS biometry
When biometry validation is called, the OS UI is displayed on top of
the app and we want to ensure that status bar and navigation bar have
the correct color matching the OS UI

Using `setFlagshipUI()` is needed only on Android because the OS UI is
only displayed between status bar and navigation bar and our app is
displayed in immersive mode. So we want to replicate the UI backdrop on
those bars

On iOS we don't have this problem as the UI backgrop already fill the
entire screen. So we want to disable this feature on iOS