fix: Set correct caret color on LockScreen on iOS
The `cursorColor` prop only works on Android so we want to use
`selectionColor` on iOS instead

The drawback of `selectionColor` is that the caret AND the selection
highlight are impacted. This is not a problem on iOS as an opacity is
applied on the highlight color. However on Android we get white text on
white highlight, which is not usable. So we want to apply
`selectionColor` only on iOS

More info:
- https://reactnative.dev/docs/textinput#cursorcolor-android
- https://reactnative.dev/docs/textinput#selectioncolor