fix: open app check if the url should be opened in an IAB
We don't want the third party dev to know about the flagship
behavior.

The dev should call openApp() when they want to open a app
even if the app will be opened on an other cozy (for instance
when sharing "cozy to cozy" a note)

So let's check if the URL of the client is the same as the
destination. If not, let's open the IAB.