feat: Create two product flavors for Android
Creating two product flavors allows to choose
the Firebase environment whatever the app build is.
It will be used to test notifications in dev or int environment
with a prod app and vice versa.

Before, app dev build was connected to dev Firebase
and app prod build was connected to prod Firebase.

- `yarn android --variant=prodDebug`
=> Firebase in "prod" and app in "dev"
- `yarn android --variant=prodRelease`
=>  Firebase in "prod" and app in "prod"
- `yarn android --variant=devDebug`
=> Firebase in "dev" and app in "dev"
- `yarn android --variant=devRelease`
=>  Firebase in "dev" and app in "prod"