refactor: Set TIMEOUT_KONNECTOR_ERROR in launcher
This constant is set to "context deadline exceeded" to be coherent with
with stack message when it forces the end of a job on timeout.

This constant is set in the launcher since clisk konnectors may be run
in other platforms than the flagship app. In this case, the same
launcher may be used across platforms