feat: Send a signal to harvest to release the execution state
When the user has run a clisk konnector, harvest display a "running"
state. If the user aborts the execution of the clisk konnector (login
failed for example), we want harvest to go back to idle state.

A signal is sent to harvest via postMessage to allow harvest to go to
the proper state.

Harvest side : https://github.com/cozy/cozy-libs/pull/2056