fix: Use div instead of react-native's button for ClouderyView mocks
Using react-native button in WebView mock would make the `render()`
method very slow on CI (>8s) which would produce a test timeout

The solution we found is to replace those button by simple divs. The
render is now very fast (a few ms), but the counterpart is that the
`fireEvent` is now slower. However the total changes are still positive
with a few seconds reduction