fix(iOS): Download file from Intent
We can display a Drive Intent from Cozy-Banks
by clicking on the chip associated to the transactions.

This Drive intent shows the Viewer and this viewer
has a "Download" button. This button was not working
because for iOS we have to make some work in order
to be download files, but we only did it for the
main frame. And since Intents are not in the main
frame, it was not working for them.

I fixed this issue by checking the link for every frames.