fix: Set appMetadata for LauncherClient
The goal is to have an instanciated cozy-client with the right metadata.
Before we had the right token, but the appMetadata was not right.

The createdByApp was right for the io.cozy.files since the stack doesn't
read the value from the client, but for triggers or bills, the
createdByApp was not right.

With this PR, we should have the right metadata now.

CozyClient is also able to take a `sourceAccount` appMetadata, but I
think we don't have a `sourceAccount` at the first launch. So we should
update it later, but I don't think cozy-client API allow this ATM.