feat: Intercept external links in ClouderyCreateInstanceView
Cloudery can display some external links in partners pages, so we
should intercept them and display them in an InAppBrowser

Those links should be opened using new tabs, so we just have to use
`onOpenWindow`. This should be the case for the "terms and condition"
PDF

We don't want to intercept navigation like we did for the ClouderyView
as the onboarding should end by redirecting to the Cozy's stack. As we
don't know the Cozy's url at this moment, then we would not be able to
distinguish it from other external URLs. So we want the Cloudery to
always use new tabs for those links