fix: Do not create new folder on error before login success
When an error was thrown in ensureAuthenticated, a folder with the name
of the account was created. This was due to the fact that no
sourceAccountIdentifier was yet fetched by the connector (and the name
of the folder is the sourceAccountIdentifier.

Now no folder is created when no sourceAccountIdentifier is available

In this case, the accountName in the account was also  rewritten to
undefined and this is also now canceled