feat: Make field autofocus more generic in login and onboarding screens
In previous implementations, we handled field autofocus from the
react-native components by predicting which field would be displayed

Those fields were easily predictable as only a few login and onboarding
scenario were implemented

Now that we implemented OIDC scenario, it is harder to predict those
fields. For example, the link from email onboarding can now redirect to
the Cloudery's FQDN choice page, or the ToS page

The new implementation allow to focus whatever DOM field that has
`autofocus` prop. If no `autofocus` field is present, then no focus is
triggered. This would prevent the keyboard to appear on the ToS page

Also as this mechanism is triggered from the HTML page and on page's
load event, then the keyboard focus is handled on every navigation,
which would make the autofocus experience more robust and homogenous