fix: Handle bar colors on TwoFactorAuthenticationView
Now that we implemented neutral screens with white background, the
TwoFactorAuthenticationView should call the `setBackgroundColor` to
re-establish the Cozy blue color

In the future may want to adapt the TwoFactorAuthenticationView to the
partner color if any. But to make this happen we should find a way to
retrieve all the theme's colors and not only the background one. Until
we find how to do this, we want the Cozy blue color to be used