fix: Declare deep links on each activities
With previous configuration, deep links were only effective on
MainActivity. If we change the default app to mespapiers, then deep
links would not work anymore. To fix that, we have to copy all
intent-filter on each activities.

We also moved all MainActivity props to activity aliases to force
MainActivitycozy to start by default. This fixes a bug when the default
activity was main activity and deep links would not work in that
scenario (because we removed them in the first step).