feat: Logout clisk webview if some account removed
Now, when running a clisk konnector, we first check that the accounts
associated to its credentials still exist in cozy database.

If some accounts in credentials do not exist, it  means that the
associated accounts have been removed from cozy database and then we can
clean the corresponding credentials.

To make this possible, credentials are now saved with the following
format :

```
{
  version: 1, // to make future migrations easier
  createdByAppVersion: '0.1.0-9999',
  createdAt: '2023-04-18T16:20:46.103Z',
  slug: 'konnectorslug', // to find credentials associated to a given konnector slug
  credentials: {
    // any object given to saveCredentials
  }
}
```

With this new credentials format, old format credentials are now
ignored.

If some credentials have been removed, it means that we must logout from
the target website of the konnector, to avoid wrong account
associations.

Konnectors must now implement a `ensureNotAuthenticated` method,
callable by the launcher. This method will be called when the some
credentials have been removed and if cozy-clisk package version is
equal greater than 0.10.0.

The launcher checks if the ensureNotAuthenticated method should be
available with the `getCliskVersion` method. In this case, the
getCliskVersion method has been merged just after the
`ensureAuthenticated` method but this is a proof of concept for the
`getCliskVersion` method.