feat: Force worker webview reload when needed
A worker webview rerender is needed when a worker url change is invoked by
the clisk konnector (goto) and the worker webview current url is already
the same url but the webview inner url is different.

By default in this case the react-native-webview component is not
rerendered si all the properties are the same.

Here we force the rerender with a change in the `key` attribute of the
webview.

Since in this case the reference to the webview is changed, we need to
update it in the launcher too.