fix: Check if icon is the same before changing it
react-native-change-icon changes icon in iOS at the
beginning of the app, even if the icon is the same as the default
one. Now we check before changing it. It was not possible to check before
because I did not configured properly activities on Android and
getIcon method was crashing. But this was fix in a previous commit.