fix: App can not be opened anymore when switching icon 2x
We have two icons and only one can be  enabled at the same time.
When we change to mespapiers icon, plugin adds cozy activity
to "classes to kill". If we change directly to cozy icon again,
plugin adds mespapiers activity to "classes to kill". So when current
activity is paused, both activities are disabled
and we can not open app anymore.