feat: Make login and onboarding views full screen
In order to allow more customization options for our onboarding
partners, we want to make the onboarding and login screens fullscreen

That will allow the Cloudery to display more complexe backgrounds
(i.e. with gradients)

This implies that status bar and navigation bar safe areas should now
be handled by the webview, so we inject js code to them to add
corresponding paddings on body element

Note that corresponding native colored bars have been removed but we
keep the `backgroundColor` state, this is because we still compute the
most effective contrast from it in order to display Android's status
bar icons colors. Also we still use it to handle transition screens
colors