fix: Resolve setWorkerState only when webview reference is ready
In some cases, we still had the following message :
"nodeHandle expected to be non-null"

This was a regression after https://github.com/cozy/cozy-flagship-app/pull/751
In the case a webview reload was forced :

Ex : when the clisk konnector does :

```javascript
await this.goto(url)
await this.runInWorker('any  command')
```

The goto command resolved when the worker webview was not ready yet and
the launcher did not have a reference to the new webview yet.

Now, the goto command waits for the webview reference to be ready to be
resolved.