feat: Add userPersistedStorage
Add a wrapper to AsyncStorage supporting storage by user. So for a UserPersistedStorageKey, we can store a value for different user.

Every keys are prefixed with @user_<user uri>_ so we can easily find every keys for an user.

It will be used in the backup feature to keep backup config when the user log out.