feat: Allow to inject credentials in the flagship app
The injection of credentials is needed when you want to test the
execution of a clisk konnector with an unauthenticated webview but with
saved credentials easily without having to wait for the end of website
session.

To do that you need to:

  - set the flag test.flagship.clisk.credentials.injecter to true
  - import the credentials in the doctype io.cozy.test.credentials with
  the following form :

```javascript
{
  "_id": "edf", // the slug of the targetted clisk konnector,
  "value": {
    "login": "login to import",
    "password": "password to import"
  }
}
```

Thi this setup, in the next run, the konnector will have the following
credentials available :

```javascript
{
  "login": "login to import",
  "password": "password to import"
}
```

When injected, the credentials will be removed from the database.