refactor: update wakeup lock check
After removing the global didRenderOnce from home,
we had some cases where a double lock could happen because the home was
rerendered due to a side-effect in globalAppState.
To remove that, we make globalAppState use
the same locking service as homeView,
with this centralized logic we remove potential conflicts.
We also add tests and handle edge cases that
were not previously handled