fix: Fix react-native-change-icon with more than two icons
In a previous patch-package, I fixed a bug in react-native-change-icon
where you could switch from icon A to B to A,  which disabled icon
A and B and led to app being impossible to start. But this fix was
only working if we had only 2 icons.

Now it works with more than 2 icons by avoiding to add
the new icon to classes to kill, instead of emptying classes
to kill each time we change icon.