fix: Do not change the user agent on pilot webview
When a clisk konnector called the `setUserAgent` function, it changed
the user agent on worker and pilot webviews.

It seems that this change of user agent on the pilot webview is causing
the pilot to loose the course of it's execution, making it wait forever.

Since now we do not download any file from the pilot itself, I think
removing the user agent from the pilot webview has no impact.