feat(backup): Split backup preparation and improve data returned to front end
prepareBackup method will be in charge
of everything needed before doing a backup :
- managing permissions
- initializing backup
- geting medias to backup
- returning data to the user

More data are returned to the front end so that the user can choose
with full knowledge of facts if he want to do a backup or not.

A backup status has been added to improve tracability for the user.