feat(backup): Add and use device id in backup folder metadata
Previously, we used the name of a remote backup folder if it
corresponds to the name of the device to identify a remote backup
corresponding to the device. Now we store a device id in remote
backup folder metadata and restore the remote backup by comparing
this device id and not the device name.

It is more robust because two devices can not have the same device
id. See react-native-device-info getUniqueId to check what is the id
we use and what are its limits.