feat: Download files progressively only when needed
Now call cozy-clisk saveFiles directly for all the given files and pass
a callback which saveFiles will call when the download of a file is
needed. This downloadAndFormatFile callback downloads the given file in
the worker webview and format it to ArrayBuffer as expected by the cozy
stack.

The index of existing files associated to the current konnector and
sourceAccountIdentifer is also created at the start of the konnector and
used on every savefiles and saveBills calls

See https://github.com/konnectors/libs/pull/949 for the related
cozy-clisk PR