feat: Intercept Manager related UniversalLinks
With current Onboarding and Login implementation, the user receives an
email containing a link to the Cloudery Manager. Then when clicked,
this link is opened into the device's Browser. Then the opened Manager
page redirects into a Universal Link that should trigger the Flagship
app with the corresponding scenario

In some environments, this scenario may not work. This is the case when
the email's link is opened into the Mailer InAppBrowser or when the OS
Browser is Firefox (by default this browser won't open external apps)

To mitigate this, we choose to convert the email's link into an
UniversalLink that has the original link into a `fallback` parameter

With this implementation, the Flagship app should be triggered whatever
is the environment, as we don't rely in a Browser anymore

However when opened by the Flagship app, we now need to open the
`fallback` parameter into a webview so the Manager can generate a new
UniversalLink (which is the same one we got in the original scenario).
Then we intercept this UniversalLink and continue the Onboarding/Login
scenario the same way we did before