fix(backup): Fix backup info when navigating in webviews with running backup
When a backup is running and we come back to photos backup page, it
runs prepareBackup which previously initialized a new backup even if
a backup was running. So now, prepareBackup return current backup
info if a backup is running, and initialize backup if not.

Removed hasLocalBackupConfig which is not used anymore.