fix: Clean IconsTable on logout
We can have different icons for the same slug depending on the
context of your cozy.

So instead of serving the wrong one, let's clean this data on
logout.

Another solution can be to store that data in the "fqdn" space,
but I don't know if it's ready or not.