feat(backup): Use nativeIntent instead of callback to send backup info to front end
On backup front end, we want to receive backup info regularly AND be able to move between webviews. Multiple webviews also need to receive backup info and display them (home and photos for the moment).

Previous callback version is not a good fit because if we start a backup on photos and we listen to backup info with a callback, the callback will be lost when we move from photos to home. It will not be recovered when we go back to photos.

By using a native intent, we can send backup info to home and photos webviews very robustly. It does not matter how or when the webview was started. If it is a home or photos webview and a backup is running, it will receive backup info regularly.