feat: Load pilot webview on about:blank
Since now the pilot webview has access to the queryAll method, the
target website may be able to request the cozy with it.

This is why we now load the pilot webview on about:blank.

Since the pilot does not handle file download anymore, it does not need
to have access to the target website.

Some clisk konnector may still run xhr requests from the pilot and they
will have to request it from the worker now.