feat(backup): Return all custom album name for an asset for iOS
@react-native-camera-roll/camera-roll was not correctly returning
album names of an asset in iOS. This patch-package fixes this, and add album
names in mediasToBackup.

In a next commit, this value will be used to add each asset to its
corresponding albums in Cozy Photos.

If and when this patch-package will be merged upstream, we will be
able to simplify this code.

Related :
https://github.com/react-native-cameraroll/react-native-cameraroll/issues/513