feat(backup): Upload Android media in appropriate folder
We decided to follow Android folder structure. So if a media is in Download, it will be uploaded in a Download folder (always inside the backup folder).

I added a new attribute for media, remotePath, corresponding to this value (like Download in my previous example).

I removed temporarly the .select in buildFilesQuery because I did not succeed in getting the path with the .select.