feat(backup): Add id to video part in photo part for Live Photo
We chose to add this id in metadata because relationships are not
possible on files and referenced_by can not remember why there is a
reference (and we want to know that is is because it is a Live
Photo).

For performance reason, this feature depends on medias to upload
order. So when there is a Live Photo, the photo part must immediately
follow the video part.

Added tests to check that.