feat: Handle Blocked Cozy on Login screen
When trying to log in a "blocked" Cozy then the app should display a
custom error page explaining to the user what is happening and what to
do

This can happen on multiple client's calls (i.e. `client.register()`)
so to be sure to intercept the error we chose to handle it in the
ErrorView which would redirect to the correct screen if detected. This
should handle classic login and OIDC, MagicLink ones

The classic login scenario depends on cozy/cozy-client#1352, otherwise
the displayed screen will be CozyNotFound

The error interception is done in two places. First on ClouderyView as
we can intercept the dedicated exception, this should be the most
common scenario. Second place is on ErrorView as a Cozy can be locked
when entering a wrong password, so we should be able to intercept this
exception anywhere in the login flow

The new CozyBlockedPage displays a forbidden logo that is a copy of the
one from cozy-stack (`assets/templates/instance_blocked.html`)

Related PR: cozy/cozy-client#1352
Related PR: cozy/cozy-client#1343