fix(backup): Set backup as to do on new start after killing app
When killing app during a backup, backup was stuck as running even if
it was not running. Now, if we start the app and backup is set as
running, we set it as to do because it is not possible. It is only
when app start, you we can still move between webview while a backup
is running