fix: Use okhttp5.0.0-alpha.1 for react-native-background-upload
We have an http2 issue when uploading a file to a cozy when we reach the
quota limit. In that case, the stack closes the connexion but it can
appears after having already send a few bytes. In that case, okhttp was
not returning the right headers and we ended up with a:
`stream closed: cancel`.

With this version, okhttp is returning the right header!

see the okhttp changelog:
https://github.com/square/okhttp/blob/master/CHANGELOG.md#version-500-alpha1
specially:
`Fix: Attempt to read the response body even if the server canceled the
request. This will cause some calls to return nice error codes like
 HTTP/1.1 429 Too Many Requests instead of transport errors like
SocketException: Connection reset and StreamResetException: stream was
reset: CANCEL.
`