fix: Better check for Upload error type in `uploadMedias()`
Previous implementation would throw if the received error do not
contain any `errors` array. this can happen if the error is not thrown
by cozy-client

Also this commit improve type check to ensure the receive error is the
one we expect. Casting objects using `as` should be avoided as this
will never ensure type compatibility